+++ backend/package.json
@@ -11,9 +11,9 @@
   "license": "ISC",
   "dependencies": {
     "@types/express": "^4.17.9",
-    "@types/multer": "^1.4.4",
-    "@types/node": "^14.14.7",
-    "@types/pg": "^7.14.6",
+    "@types/multer": "^1.4.5",
+    "@types/node": "^14.14.13",
+    "@types/pg": "^7.14.7",
     "@types/response-time": "^2.3.4",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
@@ -24,15 +24,15 @@
     "pg": "^8.5.1",
     "response-time": "^2.3.2",
     "sharp": "^0.26.3",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.0.5"
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3"
   },
   "devDependencies": {
-    "@types/cors": "^2.8.8",
+    "@types/cors": "^2.8.9",
     "@types/sharp": "^0.26.1",
-    "@typescript-eslint/eslint-plugin": "^4.8.1",
-    "@typescript-eslint/parser": "^4.8.1",
-    "eslint": "^7.13.0",
+    "@typescript-eslint/eslint-plugin": "^4.10.0",
+    "@typescript-eslint/parser": "^4.10.0",
+    "eslint": "^7.15.0",
     "eslint-plugin-react": "^7.21.5"
   }
 }
